excluindo classe cliente exemplo,cardapio,cardapioDAO, criando metodo para verificar a quantidade no estoque e alterando menu relacionado ao venda no main.
